[#12]fix : 로그인 오류 시 헬퍼테스트 동시 표시 수정
